This effectively reverts #74 without introducing a BC break: the class is now deprecated, but still exists in the codebase.
The idea of the `StubClassSourceLocator` was that it should be possible
to analyse code even if dependency holes are present, but that is now
solved by skipping BC break checks for any reflection-based exception.

This massively simplifies things, as any missing dependency will simply
lead to a skipped error and a non-zero exit code, much like it happens
in `maglnet/composer-require-checker` (https://github.com/maglnet/ComposerRequireChecker).

This also fixes weird scenarios where a pre-existing class is being
looked up in a new codebase that doesn't contain it. Assuming that
following was deleted across two breaking versions:

```php
namespace Foo;
interface Bar {}
```

Before this change, this breakage would not be located, because the
`StubClassSourceLocator` would create an implicit `Foo\Bar` stub, which
would therefore be considered as "not a change".