Prevent installation with incompatible `symfony/process` versions
Sadly, `symfony/symfony` declared `php:>=7.1` compatibility many moons ago: that is generally
technical lack of vision and responsibility on their end, but that ship has sailed.

This means that a `composer update` can and **will** install a dependency graph that is incompatible
with your system's dependencies.

This change prevents installation of `symfony/process` in versions too archaic to actually function.

As a side-effect, `vimeo/psalm` was also upgraded to its latest and greatest.

Ref: https://github.com/vimeo/psalm/releases/tag/4.12.0