#342 implemented GitHub Actions output format
This change adds a `--format=github-actions` option, allowing for friendlier output
for GitHub actions.

Note that this addition, while functionally working, is implemented by hacking through
the `MultipleChecksOn*` implementations, which decorate the various `Change` instances
produced by individual checkers. In future, we will need to clean this up, so that there
is a simpler API for creating `Change` instances from a reflected symbol.

Fixes #342