Removed redundant `true` passed to `Change::changed(..., $isBcBreak)`
`Change::changed()`, `Change::added()` and `Change::removed()` are all used when a BC break
is encountered, so the `$isBcBreak` parameter can always be omitted.

In future, we may want to rework the `Change` constructors to better encapsulate the scope of
a BC break, but for now, removing them from our own codebase already helps a lot.