Prototype: attempt to normalize `__DIR__` and `__FILE__` by changing the base path of `LocatedSource` instances
This fails due to `CompileNodeToValue()` internally using `realpath()` and `dirname()` operations, which both depend
on the current filesystem: as it currently stands, this requires an improvement of `roave/better-reflection` first.

Also, `ReplaceSourcePathOfLocatedSources` and `LocatedSourceWithStrippedSourcesDirectory` are extending `@internal`
symbols, so we may want to first make these symbols public in `roave/better-reflection`, to avoid the BC boundary to
become too relaxed, and therefore causing crashes if we (rightfully) decide to change `roave/better-reflection` internals.

Fixes #400
Fixes #410